Split main.js into new files and added docs. One test failing.
