+++ index.html
@@ -225,6 +225,9 @@
   <!-- Results output -->
   <div id="rep-comments-output" class="mt-6"></div>
 
+  <script src="js/config.js"></script>
+  <script src="js/data.js"></script>
+  <script src="js/ui.js"></script>
   <script src="js/main.js"></script>
   <script>
     const isMobile = /Mobi|Android|iPhone|iPad/i.test(navigator.userAgent);
